requirements.txt
disnake==2.7.0
python-dotenv==0.21.0
aiohttp==3.8.3
